requirements.txt
wxPython==4.2.1
mysql-connector-python==8.0.28